// dfp-piao-fe/src/app/shared/models/classes/allegato-dto.ts
import { PiaoPDFStatusEnum } from '../enums/piao-pdf-status.enum';
import { CampiTecniciDTO } from './campi-tecnici-dto';

export class AllegatoDTO extends CampiTecniciDTO {
  id?: number;
  idEntitaFK?: number;
  codDocumento?: string;
  codDocumentoFE?: string;
  codTipologiaFK?: string;
  codTipologiaAllegato?: string;
  descrizione?: string;
  downloadUrl?: string;
  sizeAllegato?: string;
  type?: string;
  isDoc?: boolean;
  base64?: string;
  status?: PiaoPDFStatusEnum;
  statusAllegato?: string;
}